Backend/requirements.txt
Django>=5.1.6
psycopg2-binary>=2.9.3
django-cors-headers>=3.10.0
python-dotenv>=0.19.2